app/requirements.txt
numpy==1.23.5
streamlit==1.30.0
tensorflow==2.15.0
Pillow==9.5.0
